requirements.txt
opencv-python
pydirectinput
pywin32
numpy
Pillow